fix some error

+++ src/app/services/services.service.ts
@@ -1,9 +1,13 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import {
+  ForbiddenException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import { CreateServiceDto } from './dto/create-service.dto';
 import { UpdateServiceDto } from './dto/update-service.dto';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Service } from './entities/service.entity';
-import { DataSource, In, Repository } from 'typeorm';
+import { DataSource, In, Not, Repository } from 'typeorm';
 import { Member } from '../members/entities/member.entity';
 
 @Injectable()
@@ -17,11 +21,7 @@ export class ServicesService {
   // create new service
   async create(createServiceDto: CreateServiceDto, orgId: number) {
     const { memberIds, ...rest } = createServiceDto;
-    const members = await this.dataSource
-      .getRepository(Member)
-      .findBy({ id: In(memberIds) });
-    if (!members) throw new NotFoundException('member not found');
-
+    const members = await this.getMembersByIds(memberIds, orgId);
     const newService = this.serviceRepository.create({
       ...rest,
       members,
@@ -29,11 +29,17 @@ export class ServicesService {
     });
     return await this.serviceRepository.save(newService);
   }
-
+  async getMembersByIds(memberIds: number[], orgId: number) {
+    const members = await this.dataSource
+      .getRepository(Member)
+      .findBy({ id: In(memberIds), orgId });
+    if (members.length !== memberIds.length)
+      throw new ForbiddenException('Some members are missing!');
+    return members;
+  }
   findAll(orgId: number) {
     return this.serviceRepository.find({
-      where: { organization: { id: orgId } },
-      relations: { members: true },
+      where: { orgId },
     });
   }
 
@@ -48,24 +54,27 @@ export class ServicesService {
     return this.serviceRepository.findBy({ id: In(ids) });
   }
 
-  async update(id: number, updateServiceDto: UpdateServiceDto) {
+  async update(id: number, updateServiceDto: UpdateServiceDto, orgId: number) {
     const { memberIds, ...rest } = updateServiceDto;
-
+    const members = await this.getMembersByIds(memberIds, orgId);
     const service = await this.serviceRepository.findOne({
       relations: { members: true },
       where: { id },
     });
-    const members = await this.dataSource
-      .getRepository(Member)
-      .findBy({ id: In(memberIds) });
-
     service.members = members;
     Object.assign(service, rest);
     return await this.serviceRepository.save(service);
   }
 
-  remove(id: number) {
-    this.serviceRepository.delete(id);
+  async remove(id: number) {
+    const service = await this.serviceRepository.findOne({
+      where: { id },
+      relations: { appointments: true },
+    });
+    if (!service) throw new NotFoundException('Service not found');
+    service.appointments = [];
+    await this.serviceRepository.save(service);
+    return this.serviceRepository.delete(id);
   }
 
   async checkOwnership(serviceId: number, orgId: number): Promise<Service> {
